Improve numerical accuracy of DRF calc at larger distances. Previous implementation used floats to perform solid-angle calculation; this would lead to numerical accuracy issues possibly starting around 12 meters! Converted to using doubles, which look to be good up to around 200 km.  I tried a few other forms of the equations, but they appeared to be less numerically accurate.

+++ src/GammaXsGui.cpp
@@ -385,8 +385,8 @@ void GammaXsGui::updateDetectorCalc()
     const string energystr = m_energyEdit->text().narrow();
     const string diststr = m_detectorDistance->text().narrow();
     const float energy = static_cast<float>( std::stod( energystr ) );
-    const float dist = static_cast<float>( PhysicalUnits::stringToDistance( diststr ) );
-    float val = m_detector->efficiency( energy, dist );
+    const double dist = PhysicalUnits::stringToDistance( diststr );
+    double val = m_detector->efficiency( energy, dist );
     char buffer[32];
     snprintf( buffer, sizeof(buffer), "%.4g", val );
     m_efficiency->setText( buffer );
